follow.js: check for invalid user input + make tabbing same as in rest of file
If a user-given character is not in specified hint following keys,
set the number id of link to follow to a very high number (32767).
So many links probably will not be visible on a page.

All the hints will be cleared from screen then.

NOTE: if, however, so many links _are_ visible, Uzbl will follow
the link with that number.

An alternative would be setting n below zero. Then hints will
remain on-screen when bad input is received.